// frontend/src/openapi/models/EvolveDialoguesMetaDto.ts
/* tslint:disable */
/* eslint-disable */

/**
 * Internal API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface EvolveDialoguesMetaDto
 */
export interface EvolveDialoguesMetaDto {
  /**
   *
   * @type {number}
   * @memberof EvolveDialoguesMetaDto
   */
  rowsNumber: number;
  /**
   *
   * @type {number}
   * @memberof EvolveDialoguesMetaDto
   */
  rowsPerPage: number;
}

export function EvolveDialoguesMetaDtoFromJSON(json: any): EvolveDialoguesMetaDto {
  return EvolveDialoguesMetaDtoFromJSONTyped(json, false)
}

export function EvolveDialoguesMetaDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): EvolveDialoguesMetaDto {
  if ((json === undefined) || (json === null)) {
    return json
  }
  return {

    'rowsNumber': json['rowsNumber'],
    'rowsPerPage': json['rowsPerPage'],
  }
}

export function EvolveDialoguesMetaDtoToJSON(value?: EvolveDialoguesMetaDto | null): any {
  if (value === undefined) {
    return undefined
  }
  if (value === null) {
    return null
  }
  return {

    'rowsNumber': value.rowsNumber,
    'rowsPerPage': value.rowsPerPage,
  }
}